pyproject.toml
[project]
name = "cityenergyanalyst"
description = "City Energy Analyst"
readme = "README.rst"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Architecture and Building Systems", email = "cea@arch.ethz.ch" }
]
maintainers = [
    { name = "Reynold Mok", email = "reynold.mok@arch.ethz.ch" }
]
dynamic = ["version"]

# Read docs/developer/dependency-management.rst for dependency management guidelines
dependencies = [
    # Scientific/geospatial packages (optimized conda builds available in conda-forge)
    # Duplicated below in [tool.pixi.dependencies] to fetch from conda-forge using pixi - update both when needed
    "numpy>=1.24.0,<2", # held back by wntr
    "scipy",
    "matplotlib",
    "geopandas>=1.0.0",
    "shapely>=2.0.0",
    "networkx>=3.4", # 3.4 contains a fix for steiner tree algorithm
    "scikit-learn",
    "numba",
    "pyyaml",
    "psutil",
    "pyarrow",
    "libpysal",
    "osmnx>=2.0.0",
    "pvlib",
    "compas",
    "compas-libigl>=0.7.6,<0.8",
    # Does not exist in pypi
    # "pythonocc-core",

    # Backend dependencies
    "aiofiles",
    "aiocache[redis]",
    "fastapi",
    "pydantic-settings",
    "uvicorn",
    "jinja2",
    "python-socketio",
    "python-multipart", # Required by fastapi for file uploads
    "sqlalchemy[asyncio]>=2.0.0",
    "sqlmodel", # Required by fastapi for database models
    "asyncpg", # Required by sqlmodel for postgres
    "aiosqlite", # Required for async sqlite adapter
    "pyjwt[crypto]", # Required by fastapi for JWT authentication

    # Other python dependencies
    "deap",
    "ephem",
    "numpy-financial",
    "python-louvain",
    "tzfpy[pytz]",
    "salib",
    "openpyxl",
    "plotly",
    "utm",
    "xlrd",
    "xlwt",
    "wntr>=1.3.2,<2", # Needs to be installed via pip since conda-forge does not provide osx-arm64 builds
    "py4design-cea",
    "pyepwmorph>=1.0.5",

    # CEA external tools
    "cea-external-tools>=0.2.1"
]

[dependency-groups]
lint = ["ruff"]
test = ["pytest", "pytest-cov"]
docs = ["setuptools", "sphinx"]
dev = [{include-group = "test"}, {include-group = "lint"}, {include-group = "docs"}]

[project.urls]
"Homepage" = "https://cityenergyanalyst.com"
"Bug Reports" = "https://github.com/architecture-building-systems/CityEnergyAnalyst/issues"
"Source" = "https://github.com/architecture-building-systems/CityEnergyAnalyst"

[project.scripts]
cea = "cea.interfaces.cli.cli:main"
cea-config = "cea.interfaces.cli.cea_config:main"
cea-doc = "cea.interfaces.cli.cea_doc:main"
cea-dev = "cea.interfaces.cli.cea_dev:main"
cea-worker = "cea.worker:main"
cea-plot = "cea.plots.plot_cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cea*"]
namespaces = false

[tool.setuptools.package-data]
cea = [
    # Configuration files
    "default.config",
    "schemas.yml",
    "scripts.yml",
    "workflows/*.yml",

    # Data files
    "examples/*.zip",
    # TODO: Move generate weather data script to scripts directory, should not be in package
    "datamanagement/weather_helper/weather.geojson",
    "databases/**/*.csv",
    "databases/weather/**/*.epw",

    # Web assets
    "interfaces/dashboard/plots/templates/*.html",

    "plots/naming.csv",
    "plots/*.html",
    "plots/**/*.png",
    "plots/**/*.yml",
]

[tool.setuptools.dynamic]
version = { attr = "cea.__version__" }

[tool.ruff]
exclude = ["*.ipynb", "external/"]
extend-exclude = ["*.pyi"]
lint.ignore = [
    "E741", # ambiguous variable name (l, O, I)
]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = [
    "linux-64", "linux-aarch64",
    "osx-64", "osx-arm64",
    "win-64"
]

[tool.pixi.tasks]
unittest = "pytest cea/tests/"
integration = "cea test --type integration"
coverage = "pytest --cov=cea cea/tests/"

# Development tasks
lint = {cmd = "ruff check --fix ."}
setup-dev = {cmd = "uv pip install -e . --group dev"}
build-wheel = {cmd = "uv build"}
export-spec = {cmd = "pixi workspace export conda-explicit-spec --ignore-pypi-errors conda-spec"}
export-conda = {cmd = "pixi workspace export conda-environment environment.yml"}

[tool.pixi.dependencies]
# Scientific/geospatial packages (also in main dependencies - conda-forge preferred)
# Duplicate of above in [project.dependencies] - update both when needed
numpy = ">=1.24.0,<2" # held back by wntr
scipy = "*"
matplotlib = "*"
geopandas = ">=1.0.0"
shapely = ">=2.0.0"
networkx = ">=3.4" # 3.4 contains a fix for steiner tree algorithm
scikit-learn = "*"
numba = "*"
pyyaml = "*"
psutil = "*"
pyarrow = "*"
pythonocc-core = "==7.8.1" # Due to #3867
pvlib = "*" # has optional C extensions for solar calculations
compas = ">=2.14.1,<3"

# Pure Python packages (conda-forge for unified dependency resolution)
libpysal = "*"
osmnx = ">=2.0.0"

# System dependencies with native libraries (fetch from conda-forge)
fiona = "*"
gdal = "*"

# Development tools (pixi-only)
notebook = "*"
pixi-pycharm = ">=0.0.8,<0.0.9" # for pycharm support
uv = "*"
timezonefinder = ">=8.1.0,<9" # to fix Windows installation due to pyepwmorpher
